// digital-library-chatbot/src/global/types/edt.ts
//edtid, title, author, year, university, program, degree, advisor, abstract, pdf, wikifier_terms
export interface EDT {
  edtid: Number;
  title: string;
  author: string;
  year: string;
  university: string;
  program: string;
  degree: string;
  advisor: string;
  abstract: string;
  pdf: string;
  wikifier_terms: string;
}